+++ test/FluentHttpRequestBuilderTest.cs
@@ -1,7 +1,10 @@
 ﻿using FluentlyHttpClient;
+using FluentlyHttpClient.Test;
+using RichardSzalay.MockHttp;
 using System.Collections.Generic;
 using System.Linq;
 using Xunit;
+using static FluentlyHttpClient.Test.ServiceTestUtil;
 using static Test.RequestBuilderTestUtil;
 
 namespace Test
@@ -152,6 +155,33 @@ namespace Test
 			Assert.Equal("hexII", chikoHeader);
 			Assert.NotNull(localeHeader);
 			Assert.Equal("mt-MT", localeHeader);
+		}
+	}
+
+	public class RequestBuilder_Return
+	{
+
+		[Fact]
+		public async void ReturnAsString()
+		{
+			var mockResponse = "{ 'name': 'Azmodan' }";
+			var mockHttp = new MockHttpMessageHandler();
+			mockHttp.When("https://sketch7.com/api/heroes/azmodan")
+				.Respond("application/json", mockResponse);
+
+			var fluentHttpClientFactory = GetNewClientFactory();
+			fluentHttpClientFactory.CreateBuilder("sketch7")
+				.WithBaseUrl("https://sketch7.com")
+				.WithMessageHandler(mockHttp)
+				.UseTimer()
+				.Register();
+
+			var httpClient = fluentHttpClientFactory.Get("sketch7");
+			var response = await httpClient.CreateRequest("/api/heroes/azmodan")
+				.ReturnAsString();
+
+			Assert.NotNull(response);
+			Assert.Equal(mockResponse, response);
 		}
 	}
 }